UIApp/requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==7.1.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
gitdb==4.0.9
GitPython==3.1.27
htbuilder==0.6.0
idna==2.10
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
iteration-utilities==0.11.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
ptyprocess==0.7.0
pyarrow==7.0.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.25.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
st-annotated-text==2.0.0
streamlit==0.84.0
terminado==0.13.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0
